requirements.txt
fastapi
uvicorn[standard]
google-cloud-speech
requests
python-dotenv
websockets
